tests/djangoherokuapp/requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.17
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
psycopg2
